move deps into partition (avoid duplication)
